Refactor: Extract voting reputation listeners into own kind
