+++ src/handlers/extensions/extensionInstalled.ts
@@ -1,4 +1,6 @@
-import { setupListenersForExtension } from '~eventListeners';
+import { Extension, getExtensionHash } from '@colony/colony-js';
+
+import { setupListenersForVotingReputation } from '~eventListeners';
 import networkClient from '~networkClient';
 import { ContractEvent } from '~types';
 import { writeExtensionFromEvent } from '~utils';
@@ -11,5 +13,8 @@ export default async (event: ContractEvent): Promise<void> => {
   );
 
   await writeExtensionFromEvent(event, extensionAddress);
-  setupListenersForExtension(extensionAddress, colony, extensionHash, false);
+
+  if (extensionHash === getExtensionHash(Extension.VotingReputation)) {
+    setupListenersForVotingReputation(extensionAddress, colony, false);
+  }
 };
